// client/src/data/types/types.gen.d.ts
import {
  type UserProfile,
  type User,
  type PortalEvent,
  type Media,
  type AirlineSettings,
  type DeliveryEntry,
  type PrintingOrderSupplier,
  type Product,
  type ExaminationSettings,
  type PriceConfiguration,
  type PrintingOrderItem,
  type PackingSlipPackage,
  type CloudFunction,
  type ApiKey,
  type Quotation,
  type ExaminationProperty,
  type AirlineOrderGrouping,
  type PriceLayer,
  type AirlineOrderGroup,
  type AirlineOrder,
  type AirlinePricing,
  type PrintingOrder,
  type PackingSlip,
  type Email,
  type BatchExport,
  type ExaminationTask,
  type Attribute,
  type AttributeOptionSet,
  type AttributeOption,
} from '../entities';

type EntityTypeName = 
 | 'UserProfile'
 | 'User'
 | 'PortalEvent'
 | 'Media'
 | 'AirlineSettings'
 | 'DeliveryEntry'
 | 'PrintingOrderSupplier'
 | 'Product'
 | 'ExaminationSettings'
 | 'PriceConfiguration'
 | 'PrintingOrderItem'
 | 'PackingSlipPackage'
 | 'CloudFunction'
 | 'ApiKey'
 | 'Quotation'
 | 'ExaminationProperty'
 | 'AirlineOrderGrouping'
 | 'PriceLayer'
 | 'AirlineOrderGroup'
 | 'AirlineOrder'
 | 'AirlinePricing'
 | 'PrintingOrder'
 | 'PackingSlip'
 | 'Email'
 | 'BatchExport'
 | 'ExaminationTask'
 | 'Attribute'
 | 'AttributeOptionSet'
 | 'AttributeOption'


type EntityName =  | 'user_profile'
 | 'user'
 | 'portal_event'
 | 'media'
 | 'airline_settings'
 | 'delivery_entry'
 | 'printing_order_supplier'
 | 'product'
 | 'examination_settings'
 | 'price_configuration'
 | 'printing_order_item'
 | 'packing_slip_package'
 | 'cloud_function'
 | 'api_key'
 | 'quotation'
 | 'examination_property'
 | 'airline_order_grouping'
 | 'price_layer'
 | 'airline_order_group'
 | 'airline_order'
 | 'airline_pricing'
 | 'printing_order'
 | 'packing_slip'
 | 'email'
 | 'batch_export'
 | 'examination_task'
 | 'attribute'
 | 'attribute_option_set'
 | 'attribute_option'


type EntityClass =
    | UserProfile
    | User
    | PortalEvent
    | Media
    | AirlineSettings
    | DeliveryEntry
    | PrintingOrderSupplier
    | Product
    | ExaminationSettings
    | PriceConfiguration
    | PrintingOrderItem
    | PackingSlipPackage
    | CloudFunction
    | ApiKey
    | Quotation
    | ExaminationProperty
    | AirlineOrderGrouping
    | PriceLayer
    | AirlineOrderGroup
    | AirlineOrder
    | AirlinePricing
    | PrintingOrder
    | PackingSlip
    | Email
    | BatchExport
    | ExaminationTask
    | Attribute
    | AttributeOptionSet
    | AttributeOption;

interface IEntityClassMap {
  UserProfile: UserProfile;
  User: User;
  PortalEvent: PortalEvent;
  Media: Media;
  AirlineSettings: AirlineSettings;
  DeliveryEntry: DeliveryEntry;
  PrintingOrderSupplier: PrintingOrderSupplier;
  Product: Product;
  ExaminationSettings: ExaminationSettings;
  PriceConfiguration: PriceConfiguration;
  PrintingOrderItem: PrintingOrderItem;
  PackingSlipPackage: PackingSlipPackage;
  CloudFunction: CloudFunction;
  ApiKey: ApiKey;
  Quotation: Quotation;
  ExaminationProperty: ExaminationProperty;
  AirlineOrderGrouping: AirlineOrderGrouping;
  PriceLayer: PriceLayer;
  AirlineOrderGroup: AirlineOrderGroup;
  AirlineOrder: AirlineOrder;
  AirlinePricing: AirlinePricing;
  PrintingOrder: PrintingOrder;
  PackingSlip: PackingSlip;
  Email: Email;
  BatchExport: BatchExport;
  ExaminationTask: ExaminationTask;
  Attribute: Attribute;
  AttributeOptionSet: AttributeOptionSet;
  AttributeOption: AttributeOption;
}